Extract indexer source identity prototype
Move RTS source-probe indexer metadata-name and body resolution into a Roslyn-backed harness identity context with direct cross-file partial indexer coverage.
